test(codegen): split the runtime-fallback proofs out of native_proof_buffer_views
The file reached 2036 lines, over the 2000-line cap. Moved the four
typed-array runtime-fallback tests into a sibling module, following the
existing native_proof_buffer_views/pointer_lifetime.rs pattern.
